[models] Actualiza el prefijo de las tablas
