+++ app/Models/Unspsc/Family.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace App\Models\Unspsc;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+
+class Family extends Model
+{
+    use HasFactory, SoftDeletes;
+
+    protected $table = 'unspsc_families';
+
+    protected $fillable = [
+        'name',
+        'code',
+        'experies_at',
+        'segment_id',
+    ];
+
+    protected $dates = [
+        'experies_at'
+    ];
+
+    public function classes()
+    {
+        return $this->hasMany(Clase::class, 'family_id')->select('id', 'name', 'code');
+    }
+
+    public function segment()
+    {
+        return $this->belongsTo(Segment::class);
+    }
+
+    public function getStatusAttribute()
+    {
+        return ($this->experies_at == null) ? 'Activo' : 'Inactivo' ;
+    }
+
+    public function getStatusColorAttribute()
+    {
+        return ($this->experies_at == null) ? 'success' : 'danger' ;
+    }
+}
