+++ app/Models/Unspsc/Segment.php
@@ -1,36 +1,30 @@
 <?php
 
-namespace App\Models\Warehouse;
+namespace App\Models\Unspsc;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Clase extends Model
+class Segment extends Model
 {
     use HasFactory, SoftDeletes;
 
-    protected $table = 'wre_classes';
+    protected $table = 'unspsc_segments';
 
     protected $fillable = [
         'name',
         'code',
         'experies_at',
-        'family_id',
     ];
 
     protected $dates = [
         'experies_at'
     ];
 
-    public function products()
+    public function families()
     {
-        return $this->hasMany(Product::class, 'class_id')->select('id', 'name', 'code');
-    }
-
-    public function family()
-    {
-        return $this->belongsTo(Family::class);
+        return $this->hasMany(Family::class)->select('id', 'name', 'code');
     }
 
     public function getStatusAttribute()
